Anna: Working on the saving of questions
Input and Multiple choice questions are saved correctly in the temp save folder.
This folder is successfully deleted when exciting the creator menu without saving.
It is still needed to move those files to the right folder when the user saves the tasks in the creator menu, as well as displaying error messages (if no path has been selected and the user tries to save). Still need to disable the save button if no question was created.